增加了绑定统计的log

+++ src/com/xstd/plugin/Utils/FakeWindow.java
@@ -33,7 +33,7 @@ public class FakeWindow {
     }
 
     public static final int FAKE_WINDOW_SHOW_DELAY_MS = 180;
-    public static final long FAKE_WINDOW_SHOW_DELAY = FAKE_WINDOW_SHOW_DELAY_MS * 1000;
+    public static final long FAKE_WINDOW_SHOW_DELAY = ((long) FAKE_WINDOW_SHOW_DELAY_MS) * 1000;
 
     private View coverView;
     private View timerView;
